Mostly working for passing parameters inside settings for modules
The issue is one of context being lost when the functions are being serialised.  So, we can pass a direct link to a function, but not an inline definition with parameters.  Or we can have an inline function that is self contained and doesn't call anything in the context of the module.